--- a/src/Microsoft.DotNet.ProjectModel/Compilation/LibraryExporter.cs
+++ b/src/Microsoft.DotNet.ProjectModel/Compilation/LibraryExporter.cs
@@ -24,8 +24,9 @@ namespace Microsoft.DotNet.ProjectModel.Compilation
         private readonly ProjectDescription _rootProject;
         private readonly string _buildBasePath;
         private readonly string _solutionRootPath;
+        private HashSet<string> _seenMetadataReferences;
 
-        private IEnumerable<LibraryExport> _cachedExports;
+        private List<LibraryExport> _cachedExports;
 
         public LibraryExporter(ProjectDescription rootProject,
             LibraryManager manager,
@@ -86,9 +87,9 @@ namespace Microsoft.DotNet.ProjectModel.Compilation
         /// </summary>
         private IEnumerable<LibraryExport> ExportLibraries(Func<LibraryDescription, bool> condition)
         {
-            _cachedExports = _cachedExports ?? CalculateAllExports().ToArray();
+            var cache = GetCacheWithRefreshedProjectExports();
 
-            foreach (var export in _cachedExports)
+            foreach (var export in cache)
             {
                 if (!condition(export.Library))
                 {
@@ -98,59 +99,80 @@ namespace Microsoft.DotNet.ProjectModel.Compilation
             }
         }
 
+        private IEnumerable<LibraryExport> GetCacheWithRefreshedProjectExports()
+        {
+            if (_cachedExports == null)
+            {
+                return (_cachedExports = CalculateAllExports().ToList());
+            }
+
+            var nonProjectExports = _cachedExports.Where(export => !Equals(export.Library.Identity.Type, LibraryType.Project));
+            var projectExports = _cachedExports.Where(export => Equals(export.Library.Identity.Type, LibraryType.Project));
+            var refreshedProjectExports = projectExports.Select(export => GenerateExportFromLibrary(_seenMetadataReferences, export.Library));
+
+            return nonProjectExports.Concat(refreshedProjectExports);
+        }
+
         private IEnumerable<LibraryExport> CalculateAllExports()
         {
-            var seenMetadataReferences = new HashSet<string>();
+            _seenMetadataReferences = new HashSet<string>();
 
             // Iterate over libraries in the library manager
             foreach (var library in LibraryManager.GetLibraries())
             {
-                var compilationAssemblies = new List<LibraryAsset>();
-                var sourceReferences = new List<LibraryAsset>();
-                var analyzerReferences = new List<AnalyzerReference>();
-                var libraryExport = GetExport(library);
-
-                // We need to filter out source references from non-root libraries,
-                // so we rebuild the library export
-                foreach (var reference in libraryExport.CompilationAssemblies)
-                {
-                    if (seenMetadataReferences.Add(reference.Name))
-                    {
-                        compilationAssemblies.Add(reference);
-                    }
-                }
-
-                // Source and analyzer references are not transitive
-                if (library.Parents.Contains(_rootProject))
-                {
-                    sourceReferences.AddRange(libraryExport.SourceReferences);
-                    analyzerReferences.AddRange(libraryExport.AnalyzerReferences);
-                }
-
-                var builder = LibraryExportBuilder.Create(library);
-                if (_runtime != null && _runtimeFallbacks != null)
-                {
-                    // For portable apps that are built with runtime trimming we replace RuntimeAssemblyGroups and NativeLibraryGroups
-                    // with single default group that contains asset specific to runtime we are trimming for
-                    // based on runtime fallback list
-                    builder.WithRuntimeAssemblyGroups(TrimAssetGroups(libraryExport.RuntimeAssemblyGroups, _runtimeFallbacks));
-                    builder.WithNativeLibraryGroups(TrimAssetGroups(libraryExport.NativeLibraryGroups, _runtimeFallbacks));
-                }
-                else
-                {
-                    builder.WithRuntimeAssemblyGroups(libraryExport.RuntimeAssemblyGroups);
-                    builder.WithNativeLibraryGroups(libraryExport.NativeLibraryGroups);
-                }
-
-                yield return builder
-                    .WithCompilationAssemblies(compilationAssemblies)
-                    .WithSourceReferences(sourceReferences)
-                    .WithRuntimeAssets(libraryExport.RuntimeAssets)
-                    .WithEmbedddedResources(libraryExport.EmbeddedResources)
-                    .WithAnalyzerReference(analyzerReferences)
-                    .WithResourceAssemblies(libraryExport.ResourceAssemblies)
-                    .Build();
+                yield return GenerateExportFromLibrary(_seenMetadataReferences, library);
             }
+        }
+
+        private LibraryExport GenerateExportFromLibrary(
+            HashSet<string> seenMetadataReferences, 
+            LibraryDescription library)
+        {
+            var compilationAssemblies = new List<LibraryAsset>();
+            var sourceReferences = new List<LibraryAsset>();
+            var analyzerReferences = new List<AnalyzerReference>();
+            var libraryExport = GetExport(library);
+
+            // We need to filter out source references from non-root libraries,
+            // so we rebuild the library export
+            foreach (var reference in libraryExport.CompilationAssemblies)
+            {
+                if (seenMetadataReferences.Add(reference.Name))
+                {
+                    compilationAssemblies.Add(reference);
+                }
+            }
+
+            // Source and analyzer references are not transitive
+            if (library.Parents.Contains(_rootProject))
+            {
+                sourceReferences.AddRange(libraryExport.SourceReferences);
+                analyzerReferences.AddRange(libraryExport.AnalyzerReferences);
+            }
+
+            var builder = LibraryExportBuilder.Create(library);
+            if (_runtime != null && _runtimeFallbacks != null)
+            {
+                // For portable apps that are built with runtime trimming we replace RuntimeAssemblyGroups and NativeLibraryGroups
+                // with single default group that contains asset specific to runtime we are trimming for
+                // based on runtime fallback list
+                builder.WithRuntimeAssemblyGroups(TrimAssetGroups(libraryExport.RuntimeAssemblyGroups, _runtimeFallbacks));
+                builder.WithNativeLibraryGroups(TrimAssetGroups(libraryExport.NativeLibraryGroups, _runtimeFallbacks));
+            }
+            else
+            {
+                builder.WithRuntimeAssemblyGroups(libraryExport.RuntimeAssemblyGroups);
+                builder.WithNativeLibraryGroups(libraryExport.NativeLibraryGroups);
+            }
+
+            return builder
+                .WithCompilationAssemblies(compilationAssemblies)
+                .WithSourceReferences(sourceReferences)
+                .WithRuntimeAssets(libraryExport.RuntimeAssets)
+                .WithEmbedddedResources(libraryExport.EmbeddedResources)
+                .WithAnalyzerReference(analyzerReferences)
+                .WithResourceAssemblies(libraryExport.ResourceAssemblies)
+                .Build();
         }
 
         private IEnumerable<LibraryAssetGroup> TrimAssetGroups(IEnumerable<LibraryAssetGroup> runtimeAssemblyGroups,