--- a/src/Microsoft.DotNet.ProjectModel/ProjectContext.cs
+++ b/src/Microsoft.DotNet.ProjectModel/ProjectContext.cs
@@ -14,7 +14,8 @@ namespace Microsoft.DotNet.ProjectModel
 {
     public class ProjectContext
     {
-        private readonly Dictionary<string, LibraryExporter> _cachedExporters = new Dictionary<string, LibraryExporter>();
+        private readonly Dictionary<Tuple<string, string>, LibraryExporter> _cachedExporters = 
+            new Dictionary<Tuple<string, string>, LibraryExporter>();
         
         private string[] _runtimeFallbacks;
 
@@ -74,7 +75,7 @@ namespace Microsoft.DotNet.ProjectModel
         public LibraryExporter CreateExporter(string configuration, string buildBasePath = null)
         {
             LibraryExporter exporter;
-            var libraryExporterCacheKey = "+ " + (configuration ?? "") + " - " + (buildBasePath ?? "");
+            var libraryExporterCacheKey = Tuple.Create(configuration ?? "" , buildBasePath ?? "");
 
             if (_cachedExporters.TryGetValue(libraryExporterCacheKey, out exporter))
             {