… src/Microsoft.DotNet.ProjectModel/ProjectModelPlatformExtensions.cs
… src/dotnet/commands/dotnet-projectmodel-server/InternalModels/ProjectContextSnapshot.cs
… test/Microsoft.DotNet.ProjectModel.Tests/GivenAProjectContext.cs
… test/Microsoft.DotNet.ProjectModel.Tests/project.json
… test/Microsoft.Extensions.DependencyModel.Tests/FunctionalTests.cs