Fixed Conflicts/Refactored/Fixed bugs on a commit from @pakrym:
https://github.com/dotnet/cli/issues/3839
https://github.com/dotnet/cli/issues/3840

Added Tests for https://github.com/dotnet/cli/issues/3839

Last commit fixes https://github.com/dotnet/cli/issues/3841
GetTypeBuildExclusion list optimized by enforcing a single visit of each node in the dependency graph, using an iterative rather than recursive graph walk, and replacing the set difference operations with a graph subset search. 

Performance increased by `.4s` on my macbook. 
Also optimized GetPlatformExclusionList graph walk by enforcing single node visit. 

Performance Data OLD:
```
Performance Summary:
[100.00% 2.392s] Thread 1
  [99.97% 99.97% 2.392s] Program:Main
    [00.28% 00.28% 0.006s] Program:ConfigureDotNetForFirstTimeUse
    [09.62% 09.62% 0.230s] Program:ResolveRootContexts
      [03.27% 33.99% 0.078s] Program:Loading project.json /Users/brthor/code/test/new/project.json
      [06.28% 65.22% 0.150s] Program:Waiting for project contexts to finish loading
    [00.13% 00.13% 0.003s] Program:Collect graph
    [88.87% 88.90% 2.126s] ProjectBuilder:Build new
      [01.43% 01.61% 0.034s] ProjectBuilder:HasSourceFiles new
      [02.21% 02.48% 0.052s] ProjectBuilder:NeedsRebuilding new
      [81.96% 92.22% 1.961s] ProjectBuilder:RunCompile new
        [19.86% 24.23% 0.475s] ProjectModelPlatformExtensions:GetTypeBuildExclusionList
          [19.84% 99.90% 0.474s] ProjectModelPlatformExtensions:GetTypeBuildExclusionList-marknonbuild
        [47.61% 58.09% 1.139s] Command:Execute dotnet /Users/brthor/code/cli/artifacts/osx.10.11-x64/stage2/sdk/1.0.0-featantares-003217/csc.dll -noconfig @/Users/brthor/code/test/new/obj/Debug/netcoreapp1.0/dotnet-compile-csc.rsp
      [01.51% 01.69% 0.036s] ProjectModelPlatformExtensions:platform exclusion
```

NEW 

```
Performance Summary:
[100.00% 1.891s] Thread 1
  [99.85% 99.85% 1.888s] Program:Main
    [00.40% 00.40% 0.007s] Program:ConfigureDotNetForFirstTimeUse
    [11.20% 11.22% 0.211s] Program:ResolveRootContexts
      [03.78% 33.78% 0.071s] Program:Loading project.json /Users/brthor/code/test/new/project.json
      [07.33% 65.47% 0.138s] Program:Waiting for project contexts to finish loading
    [00.15% 00.15% 0.002s] Program:Collect graph
    [86.56% 86.69% 1.637s] ProjectBuilder:Build new
      [01.81% 02.09% 0.034s] ProjectBuilder:HasSourceFiles new
      [03.07% 03.54% 0.057s] ProjectBuilder:NeedsRebuilding new
      [79.10% 91.38% 1.495s] ProjectBuilder:RunCompile new
        [60.08% 75.96% 1.136s] Command:Execute dotnet /Users/brthor/code/cli/artifacts/osx.10.11-x64/stage2/sdk/1.0.0-featantares-003217/csc.dll -noconfig @/Users/brthor/code/test/new/obj/Debug/netcoreapp1.0/dotnet-compile-csc.rsp
```

/cc @pakrym @anurse @eerhardt @piotrpMSFT