@brthor 2016-07-26
test RHEL7.2 x64 Release Build