@brthor 2016-08-02
Running into issues with Binding Redirects here.

The LibraryExporter cannot simply cache exported projects because the exports of a project change after it is built. Binding redirects being included in the runtime assets of that exported project is conditional on the binding redirect file existing. When the exports are cached the file doesn't exist. Post-build and pre-publish the file does exist, but the cached implementation causes it not to be copied to publish output. 

Put up a fix that side steps caching project exports, please take a look.