@brthor 2016-08-02
The changes to enable caching exports in the libraryexporter was a little more intrusive than we first thought. 

Please review @pakrym @anurse