@brthor 2016-09-07
We are taking the graph walk fixes that I made:
https://github.com/dotnet/sdk/issues/4
Closing this out though.